[FEAT] 디자인 변화에따른 UI 수정 및 파일 폴더링
