+++ momogum/Assets.xcassets/SplashLogo.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "머머금 로고_red 1.png",
+      "filename" : "SplashLogo.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
