+++ momogum/Assets.xcassets/appleSignin.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "머머금 로고_red 1.png",
+      "filename" : "appleSignin.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
